change models folders
