+++ app/models/account.rb
@@ -1,12 +1,12 @@
 class Account < ApplicationRecord
   # ASSOCIATIONS
   belongs_to :user
-  has_many :stocks, dependent: :destroy
   has_many :transactions, dependent: :destroy
   has_many :balances, dependent: :destroy
   has_many :sender_transference, class_name: 'Transference', foreign_key: 'sender_id', dependent: :destroy
   has_many :receiver_transference, class_name: 'Transference', foreign_key: 'receiver_id', dependent: :destroy
   has_many :treasuries, class_name: 'Investments::Treasury::Treasury', dependent: :destroy
+  has_many :stocks, class_name: 'Investments::Stock::Stock', dependent: :destroy
 
   monetize :balance_cents
